backend/pyproject.toml
[project]
name = "redline"
version = "0.1.0"
description = "Explainable compliance engine for HR policy verification"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "jsonschema>=4.20.0",
    "tenacity>=8.2.0",
    "pdfplumber>=0.10.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "wandb>=0.18.0",
    "weave>=0.51.0",
    "mistralai>=1.0.0",
    "huggingface-hub>=0.25.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
training = [
    "unsloth[colab-new]>=2024.8",
    "torch>=2.1.0",
    "transformers>=4.44.0",
    "datasets>=2.20.0",
    "trl>=0.9.0",
    "peft>=0.12.0",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.33.0",
]
serving = [
    "vllm>=0.5.0",
    "outlines>=0.0.46",
]
